#77 update & delete function effect

+++ app/src/main/java/com/example/stop_and_flight/RecyclerItemTouchHelper.java
@@ -12,13 +12,21 @@ import androidx.core.content.ContextCompat;
 import androidx.recyclerview.widget.ItemTouchHelper;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.example.stop_and_flight.model.Ticket;
+
 public class RecyclerItemTouchHelper extends ItemTouchHelper.SimpleCallback {
 
     private TaskAdapter adapter;
+    private TicketAdapter ticketadapter;
 
     public RecyclerItemTouchHelper(TaskAdapter adapter){
         super(0, ItemTouchHelper.LEFT | ItemTouchHelper.RIGHT);
         this.adapter = adapter;
+    }
+
+    public RecyclerItemTouchHelper(TicketAdapter adapter){
+        super(0, ItemTouchHelper.LEFT | ItemTouchHelper.RIGHT);
+        this.ticketadapter = adapter;
     }
 
     @Override
@@ -29,13 +37,27 @@ public class RecyclerItemTouchHelper extends ItemTouchHelper.SimpleCallback {
     @Override
     public void onSwiped(final RecyclerView.ViewHolder viewHolder, int direction){
         final int position = viewHolder.getAdapterPosition();
-        if(direction == ItemTouchHelper.LEFT) {
-            adapter.deleteItem(position);
-            adapter.notifyItemChanged(viewHolder.getAdapterPosition());
+        if (adapter != null)
+        {
+            if(direction == ItemTouchHelper.LEFT) {
+                adapter.deleteItem(position);
+                adapter.notifyItemChanged(viewHolder.getAdapterPosition());
+            }
+            else {
+                adapter.editItem(position);
+                adapter.notifyItemChanged(viewHolder.getAdapterPosition());
+            }
         }
-        else {
-            adapter.editItem(position);
-            adapter.notifyItemChanged(viewHolder.getAdapterPosition());
+        else if (ticketadapter != null)
+        {
+            if(direction == ItemTouchHelper.LEFT) {
+//                ticketadapter.deleteItem(position);
+                ticketadapter.notifyItemChanged(viewHolder.getAdapterPosition());
+            }
+            else {
+//                ticketadapter.editItem(position);
+                ticketadapter.notifyItemChanged(viewHolder.getAdapterPosition());
+            }
         }
     }
 
@@ -50,36 +72,74 @@ public class RecyclerItemTouchHelper extends ItemTouchHelper.SimpleCallback {
         View itemView = viewHolder.itemView;
         int backgroundCornerOffset =20;
 
-        if (dX > 0){
-            icon = ContextCompat.getDrawable(adapter.getContext(), R.drawable.ic_baseline_edit);
-            background = new ColorDrawable(ContextCompat.getColor(adapter.getContext(), R.color.colorPrimaryDark));
-        } else {
-            icon = ContextCompat.getDrawable(adapter.getContext(), R.drawable.ic_baseline_delete);
-            background = new ColorDrawable(ContextCompat.getColor(adapter.getContext(), R.color.colorPrimaryDark));
+        if (adapter != null)
+        {
+            if (dX > 0){
+                icon = ContextCompat.getDrawable(adapter.getContext(), R.drawable.ic_baseline_edit);
+                background = new ColorDrawable(ContextCompat.getColor(adapter.getContext(), R.color.colorPrimaryDark));
+            } else {
+                icon = ContextCompat.getDrawable(adapter.getContext(), R.drawable.ic_baseline_delete);
+                background = new ColorDrawable(ContextCompat.getColor(adapter.getContext(), R.color.colorPrimaryDark));
+            }
+
+            int iconMargin = (itemView.getHeight() - icon.getIntrinsicHeight()) / 2;
+            int iconTop = itemView.getTop() + (itemView.getHeight() - icon.getIntrinsicHeight()) / 2;
+            int iconBottom = iconTop + icon.getIntrinsicHeight();
+
+            if(dX > 0){
+                int iconLeft = itemView.getLeft() + iconMargin;
+                int iconRight = itemView.getLeft() + iconMargin + icon.getIntrinsicWidth();
+                icon.setBounds(iconLeft, iconTop,iconRight ,iconBottom);
+
+                background.setBounds(itemView.getLeft(), itemView.getTop(), itemView.getLeft() + ((int)dX) + backgroundCornerOffset, itemView.getBottom());
+            }
+            else if(dX < 0) {
+                int iconLeft = itemView.getRight() -  iconMargin - icon.getIntrinsicWidth();
+                int iconRight = itemView.getRight() - iconMargin;
+                icon.setBounds(iconLeft, iconTop,iconRight ,iconBottom);
+
+                background.setBounds(itemView.getRight() + ((int)dX) - backgroundCornerOffset, itemView.getTop(), itemView.getRight(), itemView.getBottom());
+            }
+            else {
+                background.setBounds(0,0,0,0);
+            }
+            background.draw(c);
+            icon.draw(c);
+        }
+        else if (ticketadapter != null)
+        {
+            if (dX > 0){
+                icon = ContextCompat.getDrawable(ticketadapter.getContext(), R.drawable.ic_baseline_edit);
+                background = new ColorDrawable(ContextCompat.getColor(ticketadapter.getContext(), R.color.colorPrimaryDark));
+            } else {
+                icon = ContextCompat.getDrawable(ticketadapter.getContext(), R.drawable.ic_baseline_delete);
+                background = new ColorDrawable(ContextCompat.getColor(ticketadapter.getContext(), R.color.colorPrimaryDark));
+            }
+
+            int iconMargin = (itemView.getHeight() - icon.getIntrinsicHeight()) / 2;
+            int iconTop = itemView.getTop() + (itemView.getHeight() - icon.getIntrinsicHeight()) / 2;
+            int iconBottom = iconTop + icon.getIntrinsicHeight();
+
+            if(dX > 0){
+                int iconLeft = itemView.getLeft() + iconMargin;
+                int iconRight = itemView.getLeft() + iconMargin + icon.getIntrinsicWidth();
+                icon.setBounds(iconLeft, iconTop,iconRight ,iconBottom);
+
+                background.setBounds(itemView.getLeft(), itemView.getTop(), itemView.getLeft() + ((int)dX) + backgroundCornerOffset, itemView.getBottom());
+            }
+            else if(dX < 0) {
+                int iconLeft = itemView.getRight() -  iconMargin - icon.getIntrinsicWidth();
+                int iconRight = itemView.getRight() - iconMargin;
+                icon.setBounds(iconLeft, iconTop,iconRight ,iconBottom);
+
+                background.setBounds(itemView.getRight() + ((int)dX) - backgroundCornerOffset, itemView.getTop(), itemView.getRight(), itemView.getBottom());
+            }
+            else {
+                background.setBounds(0,0,0,0);
+            }
+            background.draw(c);
+            icon.draw(c);
         }
 
-        int iconMargin = (itemView.getHeight() - icon.getIntrinsicHeight()) / 2;
-        int iconTop = itemView.getTop() + (itemView.getHeight() - icon.getIntrinsicHeight()) / 2;
-        int iconBottom = iconTop + icon.getIntrinsicHeight();
-
-        if(dX > 0){
-            int iconLeft = itemView.getLeft() + iconMargin;
-            int iconRight = itemView.getLeft() + iconMargin + icon.getIntrinsicWidth();
-            icon.setBounds(iconLeft, iconTop,iconRight ,iconBottom);
-
-            background.setBounds(itemView.getLeft(), itemView.getTop(), itemView.getLeft() + ((int)dX) + backgroundCornerOffset, itemView.getBottom());
-        }
-        else if(dX < 0) {
-            int iconLeft = itemView.getRight() -  iconMargin - icon.getIntrinsicWidth();
-            int iconRight = itemView.getRight() - iconMargin;
-            icon.setBounds(iconLeft, iconTop,iconRight ,iconBottom);
-
-            background.setBounds(itemView.getRight() + ((int)dX) - backgroundCornerOffset, itemView.getTop(), itemView.getRight(), itemView.getBottom());
-        }
-        else {
-            background.setBounds(0,0,0,0);
-        }
-        background.draw(c);
-        icon.draw(c);
     }
 }
